Math restructuring, scene gizmos implemented
- New cached camera type that has the same interface as the normal camera but has all the matrices stored in it
- New ECSTransformToolEx runtime resource
- Eliminated by paths for 128bit and 256bit SIMD instructions. Now, if you want to do a single operation, you can just use the low part of the 2 wide lane.
- Added PerLane SIMD operations to ease working with SIMD
- RenderingEffects now has CPU side function that transfer the GPU buffer into RAM
- Many quality of life changes for functions
- Basic Type functions extended
- New Plane and Intersection SIMD operations

+++ ECSEngine/src/ECSEngine/ECS/EntityManager.cpp
@@ -1577,7 +1577,7 @@ namespace ECSEngine {
 		memcpy(allocation, data->data, component_size);
 
 		unsigned int instance_index = manager->m_shared_components[data->component.value].instances.Add(allocation);
-		ECS_CRASH_RETURN(instance_index < ECS_SHARED_INSTANCE_MAX_VALUE, "EntityManager: Too many shared instances created for component {#}.", 
+		ECS_CRASH_RETURN(instance_index < ECS_SHARED_INSTANCE_MAX_VALUE, "EntityManager: Too many shared instances created for component {#}.",
 			manager->GetSharedComponentName(data->component));
 		
 		if (data->copy_buffers) {
@@ -4142,7 +4142,7 @@ namespace ECSEngine {
 			return false;
 		});
 
-		return { (short)instance_index };
+		return { instance_index };
 	}
 
 	// --------------------------------------------------------------------------------------------------------------------
